use tiny shall fix.

+++ doc/sql/CommentPO.sql
@@ -8,6 +8,6 @@ CREATE TABLE comment_p_o(
     `content` VARCHAR(50) NOT NULL DEFAULT '' COMMENT 'content',
     `addtime` DATETIME NOT NULL DEFAULT '1000-01-01 00:00:00' COMMENT 'addtime',
     `updatetime` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP COMMENT 'updatetime',
-    `viewed` INTEGER(12) NOT NULL DEFAULT -1 COMMENT 'viewed',
+    `viewed` tinyint NOT NULL DEFAULT -1 COMMENT 'viewed',
     PRIMARY KEY (`id`)
 )ENGINE=InnoDB DEFAULT CHARSET=utf8 COMMENT 'comment_p_o';
